run: guixsd: Update.

+++ dotfiles/run/guixsd/08-kubelet.sh
@@ -2,8 +2,12 @@
 
 set -o nounset -o errexit -o pipefail -o xtrace
 
-sudo mv /run/setuid-programs/mount.nfs /run/setuid-programs/mount.nfs.1
+if [[ -e /run/setuid-programs/mount.nfs ]]
+then
+    sudo mv /run/setuid-programs/mount.nfs /run/setuid-programs/mount.nfs.1
+fi
 
+sudo rm /var/lib/kubelet/.maintenance
 sudo herd restart kubelet
 
 mapfile -t pods < <(kubectl get -o name --namespace=openebs pods)
@@ -12,4 +16,4 @@ do
     kubectl wait --namespace=openebs --for=jsonpath='{.status.phase}'=Running --timeout=3600s "$pod"
 done
 
-sudo mv /run/setuid-programs/mount.nfs.1 /run/setuid-programs/mount.nfs
+echo sudo mv /run/setuid-programs/mount.nfs.1 /run/setuid-programs/mount.nfs
